requirements.txt
Django==4.2.0
djangorestframework==3.14.0
requests==2.31.0
celery==5.3.0
redis==4.5.5
python-decouple==3.8
django-crispy-forms==2.0
crispy-bootstrap5==0.7
plotly==5.14.1
pandas==2.0.2
numpy==1.24.3
django-extensions==3.2.3
django-cleanup==7.0.0
django-import-export==3.2.0
Pillow==10.0.0
reportlab==4.0.4
xhtml2pdf==0.2.10
python-dateutil==2.8.2
django-filter==23.3
django-widget-tweaks==1.4.12
weasyprint==60.0
django-celery-results==2.5.1
django-celery-beat==2.5.0
import-export